Segregate katello-agent tests from test_hosts
Separating all katello-agent tests from test_hosts into a new module under a new component
